pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mypkg"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "anyio==4.11.0",
    "attrs==25.3.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "lxml==6.0.2",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pluggy==1.6.0",
    "pygments==2.19.2",
    "pytest==8.4.2",
    "pytest-asyncio==0.23.7",
    "pytest-trio==0.8.0",
    "python-docx==1.2.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "trio==0.31.0",
    "typing-extensions==4.15.0",
    "pyyaml==6.0.2",
    "python-dotenv==1.0.1",
    "tqdm==4.67.1",
]

[project.optional-dependencies]
# DeepSeek-OCR CPU 환경 (macOS M1/M2, Windows 11)
# 설치: uv pip install -e ".[deepseek-cpu,dev]"
deepseek-cpu = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers==4.47.0",
    "accelerate==1.2.0",
    "qwen-vl-utils",
    "PyMuPDF==1.24.0",
    "pdf2image==1.17.0",
    "Pillow>=10.0.0",
]

# DeepSeek-OCR GPU 환경 (Windows 11 RTX 4060 / RunPod)
# 설치: uv pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# 설치: uv pip install -e ".[deepseek-gpu,dev]"
deepseek-gpu = [
    "transformers==4.47.0",
    "accelerate==1.2.0",
    "flash-attn>=2.0.0",
    "qwen-vl-utils",
    "PyMuPDF==1.24.0",
    "pdf2image==1.17.0",
    "Pillow>=10.0.0",
]

# 개발 도구
dev = [
    "jupyter==1.1.1",
    "ipykernel==6.29.5",
    "notebook==7.3.2",
    "ipywidgets==8.1.5",
]

[project.scripts]
sut-preprocess = "mypkg.cli.main:entry_point"
sut-pdf = "mypkg.cli.pdf_cli:entry_point"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mypkg = ["**/*"]  # 필요 시

[tool.pytest.ini_options]
markers = [
    "asyncio: mark a test as async"
]
